day-planner-llm/requirements.txt
fastapi
uvicorn
httpx
zeep
google-generativeai
python-dotenv
pydantic